started preparation for new design
Started preparation for new design implementation.
Due to the new design, all messages will be passed through
Dispatcher (mp-dispatcher.c).
The application code will be split into logical parts.
Each such a part named "app" / "application".
The "app" is not a really separated application, it is
a thread inside running applications.
However, the new design is agnostic to this. In the future
I can really separate the MP into separated applications,
the new is flexible enough to it.

Each "app" will implement its own functionality, like:
- APP_CONNECTION - control and status of the connection to external machines (control flow)
- APP_REMOTE - the tunnel of send and receive to/from remote machines (data flow)
- APP_CONFIG - configuration file, running configuration on this and on remote machines
- APP_PORTS - all related to ports - statuses, opening, closing, requests to firewall/router
- APP_SHELL - communication with external utils (like mpsh)
- APP_TUNNEL - tunnels control - creation, destruction, statuses, configuration
- APP_GUI - communication with GUI application (TBD)
- APP_SECURITY - certificates generation, loading, file permissions
- APP_MPFS - file system implementation (TBD)

Such a design will help to build more compact, modular and flexible
message routing.

Each "app" is a thread; each "app" will have "app_start()" and "app_stop()"
functions that will start and stop the thread internally.
Such, when we don't need an "app" we may shut it down. When we
need it again, we may start it. This way we keep resource consumption
low.